moved challenge & opp routes to separate files
