+++ src/routing/ecoverse.tsx
@@ -1,26 +1,14 @@
 import React, { FC, useMemo } from 'react';
-import { Route, Switch, useHistory, useParams, useRouteMatch } from 'react-router-dom';
+import { Route, Switch, useRouteMatch } from 'react-router-dom';
 import Loading from '../components/core/Loading/Loading';
-import {
-  useChallengeProfileQuery,
-  useChallengesQuery,
-  useOpportunityProfileQuery,
-  useOpportunityUserIdsQuery,
-} from '../hooks/generated/graphql';
+import { useChallengesQuery } from '../hooks/generated/graphql';
 import { useEcoverse } from '../hooks';
 import { useUserContext } from '../hooks';
-import { Challenge as ChallengePage, Ecoverse as EcoversePage, FourOuFour, OpportunityPage, PageProps } from '../pages';
-import {
-  AuthorizationCredential,
-  Challenge as ChallengeType,
-  ChallengesQuery,
-  Opportunity as OpportunityType,
-  User,
-} from '../models/graphql-schema';
-import ChallengeApplyRoute from './application/ChallengeApplyRoute';
+import { Ecoverse as EcoversePage, FourOuFour, PageProps } from '../pages';
+import { AuthorizationCredential } from '../models/graphql-schema';
 import { EcoverseApplyRoute } from './application/EcoverseApplyRoute';
-import { Project } from './project';
 import CommunityRoute from './community';
+import ChallengeRoute from './challenge.route';
 
 export interface RouteParameters {
   ecoverseId: string;
@@ -75,161 +63,6 @@ export const EcoverseRoute: FC<PageProps> = ({ paths }) => {
       </Route>
       <Route path={path}>
         <EcoverseApplyRoute paths={currentPaths} />
-      </Route>
-      <Route path="*">
-        <FourOuFour />
-      </Route>
-    </Switch>
-  );
-};
-
-interface ChallengeRootProps extends PageProps {
-  challenges: ChallengesQuery | undefined;
-}
-
-const ChallengeRoute: FC<ChallengeRootProps> = ({ paths, challenges }) => {
-  const { ecoverseId, toEcoverseId } = useEcoverse();
-  const { path, url } = useRouteMatch();
-  const { challengeId: id } = useParams<RouteParameters>();
-  const challengeId = challenges?.ecoverse.challenges?.find(x => x.nameID === id)?.id || '';
-
-  const { user } = useUserContext();
-
-  // todo: you don't need opportunities selected here
-  const { data: query, loading: challengeLoading } = useChallengeProfileQuery({
-    variables: {
-      ecoverseId,
-      challengeId,
-    },
-    errorPolicy: 'all',
-  });
-
-  const challenge = query?.ecoverse.challenge;
-
-  const currentPaths = useMemo(
-    () => (challenge ? [...paths, { value: url, name: challenge.displayName, real: true }] : paths),
-    [paths, id, challenge]
-  );
-
-  const isAdmin = useMemo(
-    () =>
-      user?.hasCredentials(AuthorizationCredential.GlobalAdmin) ||
-      user?.isEcoverseAdmin(toEcoverseId(ecoverseId)) ||
-      user?.isChallengeAdmin(challenge?.id || '') ||
-      false,
-    [user, ecoverseId, challenge]
-  );
-
-  const loading = challengeLoading;
-
-  if (loading) {
-    return <Loading text={'Loading challenge'} />;
-  }
-
-  if (!challenge) {
-    return <FourOuFour />;
-  }
-
-  return (
-    <Switch>
-      <Route path={`${path}/opportunities/:opportunityId`}>
-        <OpportunityRoute opportunities={challenge.opportunities} paths={currentPaths} challengeUUID={challenge.id} />
-      </Route>
-      <Route exact path={path}>
-        <ChallengePage challenge={challenge as ChallengeType} paths={currentPaths} permissions={{ edit: isAdmin }} />
-      </Route>
-      <Route path={`${path}/community`}>
-        <CommunityRoute paths={currentPaths} />
-      </Route>
-      <Route path={path}>
-        <ChallengeApplyRoute paths={currentPaths} />
-      </Route>
-      <Route path="*">
-        <FourOuFour />
-      </Route>
-    </Switch>
-  );
-};
-
-interface OpportunityRootProps extends PageProps {
-  opportunities: Pick<OpportunityType, 'id' | 'nameID'>[] | undefined;
-  challengeUUID: string;
-}
-
-const OpportunityRoute: FC<OpportunityRootProps> = ({ paths, opportunities = [], challengeUUID }) => {
-  const { path, url } = useRouteMatch();
-  const history = useHistory();
-  const { opportunityId: id, challengeId } = useParams<RouteParameters>();
-  const { user } = useUserContext();
-  const { ecoverseId, toEcoverseId } = useEcoverse();
-  const opportunityId = opportunities.find(x => x.nameID === id)?.id || '';
-
-  const { data: query, loading: opportunityLoading } = useOpportunityProfileQuery({
-    variables: { ecoverseId, opportunityId },
-    errorPolicy: 'all',
-  });
-
-  const { data: usersQuery, loading: usersLoading } = useOpportunityUserIdsQuery({
-    variables: { ecoverseId, opportunityId },
-    errorPolicy: 'all',
-  });
-
-  const opportunity = query?.ecoverse.opportunity;
-  const opportunityGroups = usersQuery?.ecoverse.opportunity;
-  const members = opportunityGroups?.community?.members;
-  const users = useMemo(() => members || [], [members]);
-
-  const currentPaths = useMemo(
-    () => (opportunity ? [...paths, { value: url, name: opportunity.displayName, real: true }] : paths),
-    [paths, id, opportunity]
-  );
-
-  const isAdmin = useMemo(
-    () =>
-      user?.hasCredentials(AuthorizationCredential.GlobalAdmin) ||
-      user?.isEcoverseAdmin(toEcoverseId(ecoverseId)) ||
-      user?.isChallengeAdmin(challengeUUID) ||
-      false,
-    [user, ecoverseId, challengeUUID]
-  );
-
-  const loading = opportunityLoading || usersLoading;
-
-  if (loading) {
-    return <Loading text={'Loading opportunity'} />;
-  }
-
-  if (!opportunity) {
-    return <FourOuFour />;
-  }
-
-  return (
-    <Switch>
-      <Route exact path={path}>
-        <OpportunityPage
-          challengeId={challengeId}
-          opportunity={opportunity as OpportunityType}
-          users={users as User[] | undefined}
-          paths={currentPaths}
-          onProjectTransition={project => {
-            history.push(`${url}/projects/${project ? project.nameID : 'new'}`);
-          }}
-          // TODO [ATS]: More generic way of controlling the UI based on user credentials must be implemented
-          permissions={{
-            edit: isAdmin,
-            projectWrite: isAdmin,
-            editAspect: user?.hasCredentials(AuthorizationCredential.GlobalAdminCommunity) || isAdmin,
-            editActorGroup: user?.hasCredentials(AuthorizationCredential.GlobalAdminCommunity) || isAdmin,
-            editActors: user?.hasCredentials(AuthorizationCredential.GlobalAdminCommunity) || isAdmin,
-            removeRelations: user?.hasCredentials(AuthorizationCredential.GlobalAdminCommunity) || isAdmin,
-          }}
-        />
-      </Route>
-      <Route path={`${path}/community`}>
-        <CommunityRoute paths={currentPaths} />
-      </Route>
-      <Route path={`${path}/projects`}>
-        <Project paths={currentPaths} projects={opportunity.projects} opportunityId={opportunity.id} />
       </Route>
       <Route path="*">
         <FourOuFour />
